Add a safe transformer to remove background fragment showing bug. Package refactoring. Seek to 0 on non page load.
