+++ app/src/main/java/com/jmilham/scrollingfeed/ui/main/MainFragment.kt
@@ -7,6 +7,7 @@ import android.view.ViewGroup
 import android.widget.Toast
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
+import androidx.viewpager2.widget.MarginPageTransformer
 import androidx.viewpager2.widget.ViewPager2
 import com.jmilham.scrollingfeed.R
 import com.jmilham.scrollingfeed.ui.helpers.VideoFragmentAdapter
@@ -32,9 +33,7 @@ class MainFragment : Fragment() {
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
         pager = view.findViewById(R.id.pager)
-        adapter = VideoFragmentAdapter(this, ArrayList())
-        pager?.adapter = adapter
-        pager?.offscreenPageLimit = 1 // this is a danger number if higher (and sometimes even at 1)
+        pager?.offscreenPageLimit = 2 // this is a danger number if higher (and sometimes even at 1)
         // BE CAUTIOUS OF ABOVE AND THE AMOUNT OF DECODERS THE DEVICE FOR THE MIME TYPE.
         // there are two limitations for how large this should be: decoders and memory allocation
         viewModel.liveVideos.observe(viewLifecycleOwner) {
@@ -46,10 +45,10 @@ class MainFragment : Fragment() {
 //                it.add(it.size / 2, JwAdvertisement("https://playertest.longtailvideo.com/vast/preroll-jw.xml")) // add an ad to the middle for test
                 adapter = VideoFragmentAdapter(this, it)
                 pager?.adapter = adapter
-                adapter.notifyDataSetChanged() // safe since only on fresh load
             }
         }
         viewModel.loadSomeVideoList()
+        pager?.setPageTransformer(MarginPageTransformer(1))
     }
 
     // needs rules for non empty? or some kind of formatting?
@@ -60,8 +59,8 @@ class MainFragment : Fragment() {
     // taking this as a chance to dump all the views inside ViewPager2
     override fun onDestroyView() {
         super.onDestroyView()
-        val size =  adapter.data.size
+        val size = adapter.data.size
         adapter.data = ArrayList()
-        adapter.notifyItemRangeRemoved(0,size)
+        adapter.notifyItemRangeRemoved(0, size)
     }
 }
